+++ test/run_all_versions.sh
@@ -88,6 +88,7 @@ SUITES=(
 	iceberg_deletes
 	iceberg_dv
 	iceberg_fdw
+	iceberg_fdw_estimate
 	iceberg_malformed
 	iceberg_name_mapping
 	iceberg_objstore